+++ include/elastic/exception.hpp
