+++ include/elastic/type.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "archive_exception.hpp"
+#include "exception.hpp"
 #include "concepts.hpp"
 #include "reflect.hpp"
