+++ test/attr.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "../include/elastic/type.hpp"
-#include "../include/elastic/binary_archive.hpp"
+#include "../include/elastic/archive.hpp"
 
 #include <boost/test/unit_test_suite.hpp>
 #include <sstream>
