+++ test/test.vcxproj
@@ -136,15 +136,17 @@
   <ItemGroup>
     <ClInclude Include="..\include\elastic.hpp" />
     <ClInclude Include="..\include\elastic\access.hpp" />
-    <ClInclude Include="..\include\elastic\archive_exception.hpp" />
-    <ClInclude Include="..\include\elastic\binary_primitive.hpp" />
+    <ClInclude Include="..\include\elastic\exception.hpp" />
+    <ClInclude Include="..\include\elastic\int128.hpp" />
+    <ClInclude Include="..\include\elastic\primitive.hpp" />
     <ClInclude Include="..\include\elastic\basic_primitive.hpp" />
-    <ClInclude Include="..\include\elastic\binary_archive.hpp" />
-    <ClInclude Include="..\include\elastic\binary_serialize.hpp" />
+    <ClInclude Include="..\include\elastic\archive.hpp" />
+    <ClInclude Include="..\include\elastic\serialize.hpp" />
     <ClInclude Include="..\include\elastic\concepts.hpp" />
-    <ClInclude Include="..\include\elastic\interface_archive.hpp" />
+    <ClInclude Include="..\include\elastic\interface.hpp" />
     <ClInclude Include="..\include\elastic\property.hpp" />
     <ClInclude Include="..\include\elastic\reflect.hpp" />
+    <ClInclude Include="..\include\elastic\type.hpp" />
     <ClInclude Include="..\include\elastic\types\float.hpp" />
     <ClInclude Include="..\include\elastic\type_traits.hpp" />
     <ClInclude Include="attr.h" />
